Implemented toml parsing; looks like i will have to remove references from structs in unit.
Implemented toml parsing using serde.
It looks like it is not possible in rust to make references as fields of struct
that reference other parts of that struct, since when you move the
value to the struct the reference becomes invalid. It might be possible
to do with some wierd updates after the struct is created but it is just
simpler to use hash maps and save key values.

+++ src/core/unit.rs
@@ -40,10 +40,11 @@ pub enum Size {
     Corps,
 }
 
+#[derive(serde::Deserialize)]
 pub struct Element {
     name: String,
     class: ElementClass,
-    cv: u32,
+    cv: f32,
     accuracy: u32,
     range: u32,
     v_inf: u32,
@@ -51,11 +52,12 @@ pub struct Element {
 }
 
 impl Element {
-    pub fn new(name: String, class: ElementClass, cv: u32, accuracy: u32, range: u32, v_inf: u32, v_arm: u32) -> Element {
+    pub fn new(name: String, class: ElementClass, cv: f32, accuracy: u32, range: u32, v_inf: u32, v_arm: u32) -> Element {
         Element { name, class, cv, accuracy, range, v_inf, v_arm }
     }
 }
 
+#[derive(serde::Deserialize)]
 pub enum ElementClass {
     Inf,
     LightTank,
